♻️ api: throw error when candidat does not exists

+++ api/src/certification/enrolment/domain/usecases/delete-unlinked-certification-candidate.js
@@ -2,6 +2,7 @@
  * @typedef {import('./index.js').CandidateRepository} candidateRepository
  */
 
+import { NotFoundError } from '../../../../shared/domain/errors.js';
 import { CertificationCandidateForbiddenDeletionError } from '../errors.js';
 
 /**
@@ -10,6 +11,10 @@ import { CertificationCandidateForbiddenDeletionError } from '../errors.js';
  */
 const deleteUnlinkedCertificationCandidate = async function ({ candidateId, candidateRepository }) {
   const candidate = await candidateRepository.get({ certificationCandidateId: candidateId });
+
+  if (!candidate) {
+    throw new NotFoundError();
+  }
 
   if (!candidate.isLinkedToAUser()) {
     return candidateRepository.remove({ id: candidateId });
